requirements.txt
pandas
numpy
requests
Pillow
